+++ src/libmongoc/tests/test-mongoc-client-side-encryption.c
@@ -6442,7 +6442,7 @@ typedef enum {
 struct kms_connect_data {
    int call_count;
    char last_host[256];
-   int32_t last_port;
+   uint16_t last_port;
    bool return_null;
    bool set_error;
    kms_proxy_transport_t transport;
@@ -6471,7 +6471,7 @@ _kms_proxy_address(kms_proxy_transport_t transport, const char **host_out, int *
  * either returns NULL with an error (forcing the KMS request to fail) or
  * connects to a hard-coded localhost proxy. */
 static mongoc_stream_t *
-_kms_connect_callback_record_and_fail(const char *host, int32_t port, void *userdata, bson_error_t *error)
+_kms_connect_callback_record_and_fail(const char *host, uint16_t port, void *userdata, bson_error_t *error)
 {
    struct kms_connect_data *data = (struct kms_connect_data *)userdata;
    data->call_count++;
@@ -6496,7 +6496,7 @@ _kms_connect_callback_record_and_fail(const char *host, int32_t port, void *user
  * itself is TLS-wrapped, and the CONNECT handshake is sent over that TLS
  * stream. */
 static mongoc_stream_t *
-_kms_connect_callback_via_proxy(const char *host, int32_t port, void *userdata, bson_error_t *error)
+_kms_connect_callback_via_proxy(const char *host, uint16_t port, void *userdata, bson_error_t *error)
 {
    struct kms_connect_data *data = (struct kms_connect_data *)userdata;
    data->call_count++;
